Feat: atualização das components navbar e footer, e da página home
