--- src/components/main/Main.jsx
@@ -1,8 +1,0 @@
-
-
-function Main (){
-    return(
-        <h1>Como funciona?</h1>
-    )
-}
-export default Main
