--- src/global/styles/index.css
@@ -1,6 +1,0 @@
-@import url("https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap");
-
-body {
-  font-family: "Montserrat", sans-serif;
-  background-color: #deedf5;
-}
