+++ src/styles/GlobalStyles.js
@@ -1,0 +1,15 @@
+import { createGlobalStyle } from "styled-components";
+
+export default createGlobalStyle`
+    *{
+        margin: 0;
+        padding: 0;
+        box-sizing: border-box;
+    }
+    @import url("https://fonts.googleapis.com/css2?family=Montserrat:ital,wght@0,100..900;1,100..900&display=swap");
+
+body {
+    font-family: "Montserrat", sans-serif;
+    background-color: #deedf5;
+}
+`
